third_party/WebKit/LayoutTests/http/tests/websocket/reserved-bits-expected.txt

CONSOLE ERROR: line 25: WebSocket connection to 'ws://localhost:8880/reserved-bits?compressed=false&bitNumber=1' failed: One or more reserved bits are on: reserved1 = 1, reserved2 = 0, reserved3 = 0
CONSOLE ERROR: line 25: WebSocket connection to 'ws://localhost:8880/reserved-bits?compressed=false&bitNumber=2' failed: One or more reserved bits are on: reserved1 = 0, reserved2 = 1, reserved3 = 0
CONSOLE ERROR: line 25: WebSocket connection to 'ws://localhost:8880/reserved-bits?compressed=false&bitNumber=3' failed: One or more reserved bits are on: reserved1 = 0, reserved2 = 0, reserved3 = 1
CONSOLE ERROR: line 25: WebSocket connection to 'ws://localhost:8880/reserved-bits?compressed=true&bitNumber=2' failed: One or more reserved bits are on: reserved1 = 0, reserved2 = 1, reserved3 = 0
CONSOLE ERROR: line 25: WebSocket connection to 'ws://localhost:8880/reserved-bits?compressed=true&bitNumber=3' failed: One or more reserved bits are on: reserved1 = 0, reserved2 = 0, reserved3 = 1
Test whether WebSocket rejects frames whose reserved bit is on.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Case compressed=false&bitNumber=1: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=false&bitNumber=2: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=false&bitNumber=3: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=true&bitNumber=2: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case compressed=true&bitNumber=3: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
